requirements.txt
streamlit>=1.20.0
ultralytics==8.3.213
torch>=2.0.0
pillow>=9.0.0
opencv-python-headless>=4.7.0
numpy>=1.21.0
